clang-tidy: fixed the issues found by 5 more checks
Applied the fixes for:
- modernize-loop-convert (the loop variables were renamed to the house style entry/rEntry
  or a descriptive name, and the asserts that still used the old iterator were fixed)
- modernize-use-starts-ends-with
- readability-avoid-return-with-void-value
- readability-redundant-casting
- readability-use-std-min-max

The fixes from these checks were reviewed and not applied:
- readability-make-member-function-const: functions that change GL/Vulkan state through a handle
  are not const even if they do not modify a member.
- cppcoreguidelines-prefer-member-initializer: clang-tidy 23 moves assignments that depend on
  earlier constructor statements into the initializer list (for example glGetUniformLocation
  before the program is created, and Plane::D before Normal is set).

+++ DemoApps/Shared/UI/Benchmark.Persistence/source/Shared/UI/Benchmark/Persistence/Input/Custom/AppInputCommandListIO.cpp
@@ -214,7 +214,7 @@ namespace Fsl::AppInputCommandListIO
       }
       if (currentVersion >= VersionId::Version3)
       {
-        timestamp = static_cast<int32_t>(ValueCompression::ReadSimpleInt32(rSrcSpan));
+        timestamp = ValueCompression::ReadSimpleInt32(rSrcSpan);
       }
       return {frameIndex, commandId, MillisecondTickCount32(timestamp), CustomWindowId(windowId), windowRectPx, mousePositionPx, isTouch};
     }
@@ -231,7 +231,7 @@ namespace Fsl::AppInputCommandListIO
       {
         commandList[i] = ReadInputCommandRecord(rSrcSpan, currentVersion);
       }
-      return {recordResolution, recordDensityDpi, std ::move(commandList), frameCount};
+      return {recordResolution, recordDensityDpi, std::move(commandList), frameCount};
     }
   }
