+++ DemoFramework/FslBase/include/FslBase/Bits/ByteSpanUtil_ReadBE.hpp
@@ -73,8 +73,8 @@ namespace Fsl::ByteSpanUtil
   constexpr inline uint32_t ReadUInt32BE(const ReadOnlySpan<uint8_t> src)
   {
     assert(src.size() >= 4);
-    return static_cast<uint32_t>(static_cast<uint32_t>(src[3]) | (static_cast<uint32_t>(src[2]) << 8) | (static_cast<uint32_t>(src[1]) << 16) |
-                                 (static_cast<uint32_t>(src[0]) << 24));
+    return (static_cast<uint32_t>(src[3]) | (static_cast<uint32_t>(src[2]) << 8) | (static_cast<uint32_t>(src[1]) << 16) |
+            (static_cast<uint32_t>(src[0]) << 24));
   }
 
   //! @brief Read a int32_t from the given index in big endian format
@@ -89,19 +89,18 @@ namespace Fsl::ByteSpanUtil
   constexpr inline uint64_t ReadUInt64BE(const ReadOnlySpan<uint8_t> src)
   {
     assert(src.size() >= 8);
-    return static_cast<uint64_t>(static_cast<uint64_t>(src[7]) | (static_cast<uint64_t>(src[6]) << 8) | (static_cast<uint64_t>(src[5]) << 16) |
-                                 (static_cast<uint64_t>(src[4]) << 24) | (static_cast<uint64_t>(src[3]) << 32) |
-                                 (static_cast<uint64_t>(src[2]) << 40) | (static_cast<uint64_t>(src[1]) << 48) |
-                                 (static_cast<uint64_t>(src[0]) << 56));
+    return (static_cast<uint64_t>(src[7]) | (static_cast<uint64_t>(src[6]) << 8) | (static_cast<uint64_t>(src[5]) << 16) |
+            (static_cast<uint64_t>(src[4]) << 24) | (static_cast<uint64_t>(src[3]) << 32) | (static_cast<uint64_t>(src[2]) << 40) |
+            (static_cast<uint64_t>(src[1]) << 48) | (static_cast<uint64_t>(src[0]) << 56));
   }
 
   //! @brief Read a int64_t from the given index in big endian format
   constexpr inline int64_t ReadInt64BE(const ReadOnlySpan<uint8_t> src)
   {
     assert(src.size() >= 8);
-    return static_cast<int64_t>(static_cast<int64_t>(src[7]) | (static_cast<int64_t>(src[6]) << 8) | (static_cast<int64_t>(src[5]) << 16) |
-                                (static_cast<int64_t>(src[4]) << 24) | (static_cast<int64_t>(src[3]) << 32) | (static_cast<int64_t>(src[2]) << 40) |
-                                (static_cast<int64_t>(src[1]) << 48) | (static_cast<int64_t>(src[0]) << 56));
+    return (static_cast<int64_t>(src[7]) | (static_cast<int64_t>(src[6]) << 8) | (static_cast<int64_t>(src[5]) << 16) |
+            (static_cast<int64_t>(src[4]) << 24) | (static_cast<int64_t>(src[3]) << 32) | (static_cast<int64_t>(src[2]) << 40) |
+            (static_cast<int64_t>(src[1]) << 48) | (static_cast<int64_t>(src[0]) << 56));
   }
 
   // -----------------------------------------------------------------------------------------------------------------------------------------------
